remote query unit test and make builder replace setters

+++ src/main/java/com/followup/davidson/dto/TjDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 public class TjDto {
     private Long tjId;
-    private double tarif;
+    private Float tarif;
     private Long personId;
     private Long projectId;
 }
